Re-add an alternative CI
OpenSUSE has no qtwebkit anymore. So pick Fedora 42, which has it and
yet appears recent. I tried to strike a balance between not being too
expensive and still being useful, because dependencies aren't part of
the image - contrary to the previous KDE-wide approach. For this reason
I haven't included Akonadi because it seems to pull in at least 1 GB of
additional dependencies.