Make Qt Webkit an optional build requirement
Disable building of everything Qt Webkit as it has been
deprecated for a long time, and there is no official Qt 6
replacement. It's getting harder to find it packaged,
and even compiling it is getting harder as it requires
other outdated stuff (hello Python 2.7).

At the same time Qt 5.15 is already 5 years old.
Most of the rest of world has moved on to Qt 6,
so it makes less and less sense to develop against
other Qt 5 based libraries.

Regardless what replaces Qt Webkit, I think it is time to
make it optional so that it is easier to hack on Trojitá.

Another benefit of this is that it allows us to build with
both Qt 5 and 6, to check for differences in unit tests,
etc between Qt versions.