Split QDataStream setting into groups and arrays
QList<T>::size() returns qsizetype in Qt 6, while it returned
an int in Qt 5. This breaks saving/reading of configs, since the
datatype is wider. Use the builtin mechanisms of QSettings for groups
and arrays instead of serializing/deserializing with QDataStream. This
leads to simpler code, and a config file that is easier to troubleshoot.
The old settings are converted on the first run, and left in place.