Fix crashing IMAP threading unit test
Check if iterator "it" is QHash<K,V>::constEnd() before trying to access
it, since constEnd(), like STL points to the imaginary object after
the last element in the QHash. Looking at the assert, "it" is expected
to be equal to QHash<K,V>::constEnd() if threadingInFlight is true,
and testing with Qt 5 reveals that "it" is equal to
QHash<K,V>::constEnd, in the same unit test:
ImapModelThreadingTest::testDynamicSearch().

Refactor the logic checking for it == constEnd() && threadingInFlight,
and continue the for loop if it == constEnd to avoid undefined behaviour.

Without this patch test_Imap_Threading crashes when built against Qt 6.9.